+++ frontend/src/App.tsx
@@ -3,7 +3,7 @@ import './App.css'
 import HomePage from "./pages/home/HomePage.tsx";
 import {BrowserRouter, Route, Routes} from "react-router-dom";
 import NewBlogEntry from "./pages/blogentry/NewBlogEntry.tsx";
-import EditBlogEntry2 from "./pages/blogentry/EditBlogEntry2.tsx";
+import EditBlogEntry from "./pages/blogentry/EditBlogEntry.tsx";
 export default function App()
 {
   return (
@@ -12,7 +12,7 @@ export default function App()
                 <Routes>
                     <Route path="/" element={<HomePage />} />
                     <Route path="/newentry" element={<NewBlogEntry />} />
-                    <Route path="/edit-entry/:id" element={<EditBlogEntry2 />} />
+                    <Route path="/edit-entry/:id" element={<EditBlogEntry />} />
                 </Routes>
             </BrowserRouter>
           </div>
